added jupyter notebook
