+++ qaoa.py
@@ -47,6 +47,26 @@ def cost_MaxCut(x,G):
         C = C + w/2*(1-(2*x[i]-1)*(2*x[j]-1))
     return C
 
+def enumerate(G):
+    if (len(G) > 30):
+        raise Exception("Too many solutions to enumerate.")
+
+    maxcut = []
+    maxcut_value = 0
+    N = len(G)
+    for i in range(2**N - 1):
+        x_bin = format(i, 'b').zfill(N)
+        x = [int(j) for j in x_bin]
+        c = 0
+        for u,v in G.edges():
+            c += G[u][v]['weight']/2*(1-(2*x[int(u)]-1)*(2*x[int(v)]-1))
+
+        if (c > maxcut_value):
+            maxcut = x
+            maxcut_value = c
+
+    return maxcut_value, maxcut
+
 def listSortedCosts_MaxCut(G):
     costs={}
     maximum=0
